# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DND AI Roleplay",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/campaigns/{campaign_id}/npcs": {
      "get": {
        "tags": [
          "campaign-npcs"
        ],
        "summary": "List Npcs",
        "operationId": "list_npcs_api_v1_campaigns__campaign_id__npcs_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NpcView"
                  },
                  "title": "Response List Npcs Api V1 Campaigns  Campaign Id  Npcs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "campaign-npcs"
        ],
        "summary": "Create Npc",
        "operationId": "create_npc_api_v1_campaigns__campaign_id__npcs_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NpcCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/npcs:extract": {
      "post": {
        "tags": [
          "campaign-npcs"
        ],
        "summary": "Extract Npcs",
        "operationId": "extract_npcs_api_v1_campaigns__campaign_id__npcs_extract_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NpcView"
                  },
                  "title": "Response Extract Npcs Api V1 Campaigns  Campaign Id  Npcs Extract Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/npcs/{npc_id}": {
      "patch": {
        "tags": [
          "campaign-npcs"
        ],
        "summary": "Update Npc",
        "operationId": "update_npc_api_v1_npcs__npc_id__patch",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Npc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NpcUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "campaign-npcs"
        ],
        "summary": "Delete Npc",
        "operationId": "delete_npc_api_v1_npcs__npc_id__delete",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Npc Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NpcCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "role": {
            "type": "string",
            "maxLength": 400,
            "title": "Role",
            "default": ""
          },
          "personality": {
            "type": "string",
            "maxLength": 800,
            "title": "Personality",
            "default": ""
          },
          "ideal": {
            "type": "string",
            "maxLength": 400,
            "title": "Ideal",
            "default": ""
          },
          "bond": {
            "type": "string",
            "maxLength": 400,
            "title": "Bond",
            "default": ""
          },
          "flaw": {
            "type": "string",
            "maxLength": 400,
            "title": "Flaw",
            "default": ""
          },
          "knows": {
            "type": "string",
            "maxLength": 1200,
            "title": "Knows",
            "default": ""
          },
          "wants": {
            "type": "string",
            "maxLength": 600,
            "title": "Wants",
            "default": ""
          },
          "voice": {
            "type": "string",
            "maxLength": 600,
            "title": "Voice",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "NpcCreate"
      },
      "NpcUpdate": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "role": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 400
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "personality": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 800
              },
              {
                "type": "null"
              }
            ],
            "title": "Personality"
          },
          "ideal": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 400
              },
              {
                "type": "null"
              }
            ],
            "title": "Ideal"
          },
          "bond": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 400
              },
              {
                "type": "null"
              }
            ],
            "title": "Bond"
          },
          "flaw": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 400
              },
              {
                "type": "null"
              }
            ],
            "title": "Flaw"
          },
          "knows": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1200
              },
              {
                "type": "null"
              }
            ],
            "title": "Knows"
          },
          "wants": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 600
              },
              {
                "type": "null"
              }
            ],
            "title": "Wants"
          },
          "voice": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 600
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          }
        },
        "type": "object",
        "required": [
          "revision"
        ],
        "title": "NpcUpdate"
      },
      "NpcView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "campaignId": {
            "type": "string",
            "title": "Campaignid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "personality": {
            "type": "string",
            "title": "Personality"
          },
          "ideal": {
            "type": "string",
            "title": "Ideal"
          },
          "bond": {
            "type": "string",
            "title": "Bond"
          },
          "flaw": {
            "type": "string",
            "title": "Flaw"
          },
          "knows": {
            "type": "string",
            "title": "Knows"
          },
          "wants": {
            "type": "string",
            "title": "Wants"
          },
          "voice": {
            "type": "string",
            "title": "Voice"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "campaignId",
          "name",
          "role",
          "personality",
          "ideal",
          "bond",
          "flaw",
          "knows",
          "wants",
          "voice",
          "source",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "title": "NpcView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}